+++ components/category/CardListContainer/index.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { useRecoilValue } from "recoil";
 import { lectureDataList } from "store/state";
+import { LectureDataListType, ResponseLectureDataType } from "types/info.type";
 
 import CardBox from "../CardBox";
 import { StyledRoot } from "./style";
@@ -10,7 +11,7 @@ function CardListContainer() {
 
   return (
     <StyledRoot>
-      {LectureDataList?.data.map((lecture) => (
+      {LectureDataList?.map((lecture: ResponseLectureDataType) => (
         <CardBox key={lecture.LectureTitle} lecture={lecture} />
       ))}
       {/* <CardBox />
